Rename shaders/scene to shaders/material.
