+++ src/framework/shaders/material/pbr_metallic_roughness/scene.frag.glsl
@@ -5,9 +5,8 @@
 
 // ----------------------------------------------------------------------------
 
-#include "scene/interop.h"
-#include "scene/pbr_metallic_roughness/interop.h"
-
+#include <material/interop.h>
+#include <material/pbr_metallic_roughness/interop.h>
 #include <shared/maths.glsl>
 #include <shared/lighting/pbr.glsl>
